+++ server/vitals.js
@@ -147,62 +147,35 @@ async function checkButtonBatteries() {
 // - reconnectedButtons: An array of button display names that refer to buttons that have recently reconnected
 function sendClientButtonStatusChanges(clientButtonStatusChanges) {
   // Loop through each client to create the Twilio messages
-  Object.values(clientButtonStatusChanges).forEach(({ client, disconnectedButtons, reconnectedButtons }) => {
-    let message = [i18next.t('buttonStatusChangeStart', { lng: client.language, clientDisplayName: client.displayName })]
-    let messageBilingual = null
+  Object.values(clientButtonStatusChanges).forEach(buttonStatusChanges => {
+    const disconnectedButtons =
+      buttonStatusChanges.disconnectedButtons.length > 0 ? buttonStatusChanges.disconnectedButtons.sort().join(', ') : undefined
+    const reconnectedButtons =
+      buttonStatusChanges.reconnectedButtons.length > 0 ? buttonStatusChanges.reconnectedButtons.sort().join(', ') : undefined
+    let translation
 
-    // check for bilingual client language
-    if (message[0].split('\n---\n').length > 1) {
-      const split = message[0].split('\n---\n')
-
-      message = [split[0]]
-      messageBilingual = [split[1]]
+    if (disconnectedButtons !== undefined && reconnectedButtons !== undefined) {
+      translation = 'buttonStatusChangeDisconnectedAndReconnected'
+    } else if (disconnectedButtons !== undefined) {
+      translation = 'buttonStatusChangeDisconnected'
+    } else {
+      // NOTE: all client button status changes must have either disconnected or reconnected buttons
+      // therefore, if there are no disconnected buttons, there mustbe reconnected buttons
+      translation = 'buttonStatusChangeReconnected'
     }
 
-    if (disconnectedButtons.length > 0) {
-      const buttonDisplayNames = disconnectedButtons.sort().join(', ') // sorted alphabetically
-      let disconnectButtonsMessage = i18next.t('buttonStatusChangeDisconnected', { lng: client.language, buttonDisplayNames })
-
-      // handle bilingual client language translation
-      if (messageBilingual !== null) {
-        const split = disconnectButtonsMessage.split('\n---\n')
-
-        disconnectButtonsMessage = split[0]
-        messageBilingual.push(split[1])
-      }
-
-      message.push(disconnectButtonsMessage)
-    }
-
-    if (reconnectedButtons.length > 0) {
-      const buttonDisplayNames = reconnectedButtons.sort().join(', ') // sorted alphabetically
-      let reconnectedButtonsMessage = i18next.t('buttonStatusChangeReconnected', { lng: client.language, buttonDisplayNames })
-
-      // handle bilingual client language translation
-      if (messageBilingual !== null) {
-        const split = reconnectedButtonsMessage.split('\n---\n')
-
-        reconnectedButtonsMessage = split[0]
-        messageBilingual.push(split[1])
-      }
-
-      message.push(reconnectedButtonsMessage)
-    }
-
-    // join the message parts with spaces
-    message = message.join(' ')
-
-    // handle bilingual client language translation
-    if (messageBilingual !== null) {
-      messageBilingual = messageBilingual.join(' ')
-      message += `\n---\n${messageBilingual}`
-    }
+    const message = i18next.t(translation, {
+      lng: buttonStatusChanges.client.language,
+      clientDisplayName: buttonStatusChanges.client.displayName,
+      disconnectedButtons,
+      reconnectedButtons,
+    })
 
     // send the button status changes to the client's heartbeat and responder phone numbers
-    const recipients = [...client.heartbeatPhoneNumbers, ...client.responderPhoneNumbers]
+    const recipients = [...buttonStatusChanges.client.heartbeatPhoneNumbers, ...buttonStatusChanges.client.responderPhoneNumbers]
 
     recipients.forEach(phoneNumber => {
-      twilioHelpers.sendTwilioMessage(phoneNumber, client.fromPhoneNumber, message)
+      twilioHelpers.sendTwilioMessage(phoneNumber, buttonStatusChanges.client.fromPhoneNumber, message)
     })
   })
 }